frontend/requirements.txt
streamlit 
openai 
pyyaml
python-dotenv
requests
Pillow